# web-navigation-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web Navigation API",
    "version": "1.0.0",
    "description": "AI-powered web navigation API for autonomous agents. Navigate URLs, extract structured data, perform live search, follow link graphs, run adaptive crawls, handle anti-bot challenges, diff page content, and gate navigation actions.",
    "x-agent-callable": true,
    "x-mcp-compatible": true,
    "x-pricing": {
      "free_tier": {
        "requests_per_day": 100,
        "requests_per_month": 3000
      },
      "pay_per_call": {
        "navigate": 0.005,
        "extract-structured-data": 0.006,
        "live-search": 0.005,
        "follow-links": 0.006,
        "adaptive-crawl": 0.008,
        "anti-bot-handling": 0.004,
        "content-diff": 0.005,
        "browser-session": 0.003,
        "screenshot": 0.003,
        "form-detect": 0.004,
        "submit-form-gated": 0.005,
        "download-assets": 0.004,
        "navigate-and-extract": 0.015,
        "execution-gate": 0.002
      }
    }
  },
  "servers": [
    {
      "url": "https://orbis-apis.onrender.com/web-navigation"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/navigate": {
      "post": {
        "operationId": "navigate",
        "summary": "Navigate a URL with a goal and extract structured page intelligence",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url",
                  "goal"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "goal": {
                    "type": "string"
                  },
                  "extract": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "follow_redirects": {
                    "type": "boolean"
                  },
                  "wait_for": {
                    "type": "string",
                    "enum": [
                      "load",
                      "networkidle",
                      "domcontentloaded"
                    ]
                  },
                  "render_js": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Navigation result with page intelligence",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "final_url": {
                      "type": "string"
                    },
                    "goal": {
                      "type": "string"
                    },
                    "page_type": {
                      "type": "string"
                    },
                    "page_title": {
                      "type": "string"
                    },
                    "page_summary": {
                      "type": "string"
                    },
                    "navigation_status": {
                      "type": "string",
                      "enum": [
                        "success",
                        "redirect",
                        "blocked",
                        "error"
                      ]
                    },
                    "content_extracted": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "links_found": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          },
                          "text": {
                            "type": "string"
                          },
                          "relevance_score": {
                            "type": "number",
                            "minimum": 0,
                            "maximum": 1
                          }
                        }
                      }
                    },
                    "forms_found": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "action": {
                            "type": "string"
                          },
                          "fields": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "render_required": {
                      "type": "boolean"
                    },
                    "next_navigation_options": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "confidence_per_section": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    },
                    "recommended_actions_priority_order": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "privacy": {
                      "type": "object",
                      "properties": {
                        "data_stored": {
                          "type": "boolean"
                        },
                        "retention": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing url or goal"
          }
        }
      }
    },
    "/extract-structured-data": {
      "post": {
        "operationId": "extractStructuredData",
        "summary": "Extract structured data from a page according to a target schema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url",
                  "schema",
                  "page_content"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "schema": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "page_content": {
                    "type": "string"
                  },
                  "extraction_method": {
                    "type": "string",
                    "enum": [
                      "ai",
                      "css",
                      "xpath",
                      "regex"
                    ]
                  },
                  "pagination": {
                    "type": "boolean"
                  },
                  "max_pages": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured extraction result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "extracted_data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "extraction_confidence": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "fields_extracted": {
                      "type": "number"
                    },
                    "fields_missing": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "pagination_detected": {
                      "type": "boolean"
                    },
                    "total_pages_estimate": {
                      "type": "number"
                    },
                    "data_quality": {
                      "type": "object",
                      "properties": {
                        "completeness": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 1
                        },
                        "accuracy_signals": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "anomalies": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "extraction_method_used": {
                      "type": "string"
                    },
                    "confidence_per_section": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "number"
                      }
                    },
                    "recommended_actions_priority_order": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "privacy": {
                      "type": "object",
                      "properties": {
                        "data_stored": {
                          "type": "boolean"
                        },
                        "retention": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing url, schema, or page_content"
          }
        }
      }
    },
...
  }
}